Change page subject from Bread to CSS
I started this project thinking of using the Wikipedia page about bread as a subject, but this proved to be more of a challenge than necessary, due to the excessive text on the page and the lack of necessary items in the user cases. So I decided to use the basic CSS documentation for this exercise.